requirements.txt
requests
beautifulsoup4
pandas
spacy
python-dotenv